+++ lib/generator/biomegrid/ErosionMapLayer.h
@@ -1,7 +1,6 @@
 #ifndef EXT_NOISELIB_LIB_GENERATOR_BIOMEGRID_EROSIONMAPLAYER_H_
 #define EXT_NOISELIB_LIB_GENERATOR_BIOMEGRID_EROSIONMAPLAYER_H_
 
-
 #include <utility>
 #include "MapLayer.h"
 
@@ -9,7 +8,7 @@ using namespace GridBiome;
 
 class ErosionMapLayer : public MapLayer {
  public:
-  ErosionMapLayer(int_fast64_t seed, std::shared_ptr<MapLayer> belowLayer) : MapLayer(seed), below_layer_(std::move(belowLayer)) {}
+  ErosionMapLayer(int_fast32_t seed, std::shared_ptr<MapLayer> belowLayer) : MapLayer(seed), below_layer_(std::move(belowLayer)) {}
 
   ~ErosionMapLayer();
