+++ lib/generator/decorators/DeadBushDecorator.h
@@ -5,7 +5,7 @@
 
 class DeadBushDecorator : public Decorator{
  public:
-  void Decorate(ChunkManager &world, Random &random, int_fast64_t chunkX, int_fast64_t chunkZ) override;
+  void Decorate(ChunkManager &world, Random &random, int_fast32_t chunkX, int_fast32_t chunkZ) override;
 };
 
 #endif //EXT_NOISELIB_LIB_GENERATOR_DECORATORS_DEADBUSHDECORATOR_H_
