+++ lib/generator/decorators/Decorator.cpp
@@ -4,7 +4,7 @@ void Decorator::SetAmount(int amount) {
   amount_ = amount;
 }
 
-void Decorator::Populate(ChunkManager &world, Random &random, int_fast64_t chunkX, int_fast64_t chunkZ) {
+void Decorator::Populate(ChunkManager &world, Random &random, int_fast32_t chunkX, int_fast32_t chunkZ) {
   for (int i = 0; i < amount_; ++i) {
     Decorate(world, random, chunkX, chunkZ);
   }
