+++ lib/generator/decorators/LakeDecorator.h
@@ -9,14 +9,14 @@ class LakeDecorator : public Decorator {
   LakeDecorator(MinecraftBlock mcBlock, int populatorRarity, int offset = 0)
       : block(mcBlock), rarity(populatorRarity), baseOffset(offset) {}
 
-  void Decorate(ChunkManager &world, Random &random, int_fast64_t chunkX, int_fast64_t chunkZ) override;
+  void Decorate(ChunkManager &world, Random &random, int_fast32_t chunkX, int_fast32_t chunkZ) override;
 
  private:
   /**
    * Verifies if the coordinates has no water block intersecting to this decoration object.
    * If so, it will automatically cancels this decorator object.
    */
-  static bool VerifyNoWaterBlocks(ChunkManager &manager, int_fast64_t x, int_fast32_t y, int_fast64_t z) {
+  static bool VerifyNoWaterBlocks(ChunkManager &manager, int_fast32_t x, int_fast32_t y, int_fast32_t z) {
     return !IS_LIQUID(manager.GetBlockAt(x, y, z).GetId());
   }
