+++ lib/generator/decorators/TreeDecorator.cpp
@@ -21,11 +21,11 @@ TreeObject TreeDecorator::GetRandomTree(Random random) {
   return nullptr;
 }
 
-void TreeDecorator::Decorate(ChunkManager &world, Random &random, int_fast64_t chunkX, int_fast64_t chunkZ) {
+void TreeDecorator::Decorate(ChunkManager &world, Random &random, int_fast32_t chunkX, int_fast32_t chunkZ) {
   auto chunk = world.GetChunk(chunkX, chunkZ);
 
-  int_fast64_t x = random.NextInt(16);
-  int_fast64_t z = random.NextInt(16);
+  int_fast32_t x = random.NextInt(16);
+  int_fast32_t z = random.NextInt(16);
   int_fast32_t source_y = chunk->GetHighestBlockAt(x, z);
 
   TreeObject treeObject = GetRandomTree(random);
@@ -39,7 +39,7 @@ void TreeDecorator::Decorate(ChunkManager &world, Random &random, int_fast64_t c
   }
 }
 
-void TreeDecorator::Populate(ChunkManager &world, Random &random, int_fast64_t chunkX, int_fast64_t chunkZ) {
+void TreeDecorator::Populate(ChunkManager &world, Random &random, int_fast32_t chunkX, int_fast32_t chunkZ) {
   int treeAmount = amount_;
   if (random.NextInt(10) == 0) {
     ++treeAmount;
