+++ lib/generator/ground/GroundGenerator.h
@@ -9,8 +9,8 @@ class GroundGenerator {
  public:
   virtual void GenerateTerrainColumn(ChunkManager &world,
                                      Random &random,
-                                     int_fast64_t x,
-                                     int_fast64_t z,
+                                     int_fast32_t x,
+                                     int_fast32_t z,
                                      int biome,
                                      double surface_noise);
  protected:
